+++ test/widget_test.dart
@@ -537,6 +537,9 @@ class _FakeApi implements RemoteStorageGateway {
   Future<void> deleteSyncProfile(String id) async {}
 
   @override
+  Future<void> setLogLevel(String level) async {}
+
+  @override
   Future<void> writeAppLog(
     String message, {
     String level = 'info',